Update functions.php
A constant has been created to store the Boostrap's version string and it was used in lines 97 and 103 to correct the wrong Bootrstrap version ('3.3.7').
It is just a suggestion. This variable change whenever you update the Bootstrap's version, so it is a good practice to store this value in a  constant.